Day 9 part 2 successful
Honestly I think answer-variable-type is my arch-enemy. Seriously how
many times did I f**k that up?
Also mergin free space blocks was not necessary, I did do it and the
code seemed to work, so I'm still proud of that, but it gave the same
answer, so I did not include it in the solution.