Day 13 part 1
IDK, had to look at the subreddit for inspiration on this one. I'm not
sure I entirely understands why this was the proper way; it turns out
this gives you the minimum amount of tokens because the system only has
one solution. It seems there aren't any instances where the vectors
(ax,ay) and (bx,by) are multiples of each other too.