Day 15 part 2
Honestly this was pretty fun, I like doing sokoban games.
Ended up doing it with 2 grids instead of worrying about movement order,
it's just so much simpler this way. At the start I wanted to try having
2 pointers and 2 arrays but not only I didn't manage to get it right,
but also it was unnecessary, since at the end of each movement the 2
grids stay the same, so I don't need to switch between them for the next
movement.