Day 17 part 2
BROOOOO this was amazingggg.
This code is input specific to what was given to me, but thanks to being
that specific it runs in <= 1ms so I'm really happy. Was pretty fun
simplifying the program as much as possible to make it only dependent on
A, and using that knowledge to construct the value of A that it would
need to have. I imagine there are inputs where this approach would not
completely work, since it could happen that you choose a "next_bits"
that makes it impossible to generate some of the next goals, but also it
might be impossible for that to happen since there seems to always be a
combination of bits that does, no matter the previous chosen bits.