Day 19 part 2
Quite easy adapting the part 1 solution to this. Only needed to change
bool to long and incremented solvable instead of just setting it to
true. Not quite as fast as part 1 (this was ~312ms). I'll look in the
subreddit for faster solutions just for curiosity.

Lesson learned: remember that memoization exists lol.