Small improvement to Day 19 part 2
Just replaced map for array in memo, shaved off like 10ms, so not that
much. One solution that I saw in the subreddit which was cool was using
a tree, which is something that I actually considered but just could not
come up with a way to do it. To be fair more code is needed, when really
this solution is good enough for me.