Day 24 part 2!!!
Okay so the code doesn't actually give you the answer, I just used it as
a tool to see in what places it was wrong. I ended up having to
understand the structure of the input, write down the circuit and find
what wires were wrong, switch them manually and see the amount of wrong
answers go down. Basically just understand how the circuit was
structured when it was correct and find the same pattern in the parts
that were wrong. I wonder what the more elegant solutions are like, this
problem was small enough that it could be done manually but there
could've been way more pairs of wires switched up, I'll look in the
subreddit after I'm done with the rest of the days (if I'm able to solve
Day 23 without looking at clues)